+++ pdf2json.d.ts
@@ -1,0 +1,10 @@
+declare module "pdf2json" {
+  import { EventEmitter } from "events";
+
+  export default class PDFParser extends EventEmitter {
+    constructor(context?: any, needRawText?: number);
+    parseBuffer(buffer: Buffer): void;
+    getRawTextContent(): string;
+    loadPDF(pdfFilePath: string): void;
+  }
+}